Travelista/requirements/base.txt
Django>4.2,<4.3
python-dotenv
pillow
django-taggit
